Modularize Bitergia code related to the new menu
